Create README.md
This repository contains of a bank marketing campaign, though the data is cleaned it did come with some challenges cleaning it. numerous columns that did not add to the advancement of our analysis were deleted or merged.the dataset after being and analysed provided some insights that blue collar jobs had the most representation with the most duration averaging 270secs approximately almost 4mins. the second observation could be that the month of May had the traction across all occupations. the analysis could save a as baseline for a machine learning design as regards the best possible time to make a marketing campaiggn and also which specific demographic to focus on.